Added Banking Sim Project
o	This Project is a simulation of current credit card systems. It emulates the transaction process with a credit card balance and limit just as if one was reviewing their banking app. The project uses randomized items, prices, and credit cards to simulate an individual’s day to day credit card use. Written in the IntelliJ IDE, this was for a Programming 1 class project.